+++ src/routes/auth/auth.routes.ts
@@ -3,18 +3,16 @@
  */
 
 import { createRoute, z } from '@hono/zod-openapi'
+import {
+  basicMessageResponseSchema,
+  errorResponseSchema,
+  loginBodySchema,
+  loginResponseSchema,
+  meDataSchema,
+  unauthorizedResponseSchema,
+} from '@/lib/zod-schemas'
 import jsonContent, { jsonContentRequired } from '@/middleware/utils/json-content'
 import * as httpStatusCodes from '@/openapi/http-status-codes'
-
-const loginBodySchema = z.object({
-  username: z.string().min(1),
-  password: z.string().min(1),
-})
-
-const meDataSchema = z.object({
-  sub: z.string(),
-  role: z.string(),
-})
 
 export const loginRoute = createRoute({
   tags: ['Auth'],
@@ -28,29 +26,19 @@ export const loginRoute = createRoute({
   },
   responses: {
     [httpStatusCodes.BAD_REQUEST]: jsonContent(
-      z.object({ message: z.string() }),
+      basicMessageResponseSchema,
       'Invalid request body provided',
     ),
     [httpStatusCodes.OK]: jsonContent(
-      z.object({
-        message: z.string(),
-        data: z.object({
-          id: z.string(),
-          username: z.string(),
-          role: z.string(),
-        }),
-      }),
+      loginResponseSchema,
       'Login successful',
     ),
     [httpStatusCodes.UNAUTHORIZED]: jsonContent(
-      z.object({ message: z.string() }),
+      unauthorizedResponseSchema,
       'Invalid Credentials',
     ),
     [httpStatusCodes.INTERNAL_SERVER_ERROR]: jsonContent(
-      z.object({
-        message: z.string(),
-        errors: z.any(),
-      }),
+      errorResponseSchema,
       'Internal Server Error',
     ),
   },
@@ -69,14 +57,11 @@ export const getCurrentUserRoute = createRoute({
       'Successfully retrieved user information from token',
     ),
     [httpStatusCodes.UNAUTHORIZED]: jsonContent(
-      z.object({ message: z.string() }),
+      unauthorizedResponseSchema,
       'Unauthorized. Invalid or missing token',
     ),
     [httpStatusCodes.INTERNAL_SERVER_ERROR]: jsonContent(
-      z.object({
-        message: z.string(),
-        errors: z.any(),
-      }),
+      errorResponseSchema,
       'Internal Server Error',
     ),
   },
@@ -88,18 +73,15 @@ export const logoutRoute = createRoute({
   path: '/logout',
   responses: {
     [httpStatusCodes.OK]: jsonContent(
-      z.object({ message: z.string() }),
+      basicMessageResponseSchema,
       'Logout successful',
     ),
     [httpStatusCodes.UNAUTHORIZED]: jsonContent(
-      z.object({ message: z.string() }),
+      unauthorizedResponseSchema,
       'Unauthorized. Invalid or missing token',
     ),
     [httpStatusCodes.INTERNAL_SERVER_ERROR]: jsonContent(
-      z.object({
-        message: z.string(),
-        errors: z.any(),
-      }),
+      errorResponseSchema,
       'Internal Server Error',
     ),
   },
@@ -112,18 +94,15 @@ export const refreshRoute = createRoute({
   description: 'Refreshes the access token using the refresh token',
   responses: {
     [httpStatusCodes.OK]: jsonContent(
-      z.object({ message: z.string() }),
+      basicMessageResponseSchema,
       'Access token refreshed successfully. New token is set in an httpOnly cookie.',
     ),
     [httpStatusCodes.UNAUTHORIZED]: jsonContent(
-      z.object({ message: z.string() }),
+      unauthorizedResponseSchema,
       'Unauthorized. The refresh token is missing, invalid, or expired.',
     ),
     [httpStatusCodes.INTERNAL_SERVER_ERROR]: jsonContent(
-      z.object({
-        message: z.string(),
-        errors: z.any(),
-      }),
+      errorResponseSchema,
       'Internal Server Error',
     ),
   },
